Reuse the Central search panel instead of its logic
The Add Trigger popup needed the gallery's Central results, and lifting the
search into a hook rewrote a component that has nothing to do with agents.
The only thing that differs is what a card click does, so the panel takes an
optional onSelect and the popup passes one.
