+++ packages/ballerina-visualizer/src/views/BI/ComponentListView/CentralSearchPanel.tsx
@@ -15,9 +15,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import React from 'react';
+import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { useRpcContext } from '@wso2/ballerina-rpc-client';
 import { EVENT_TYPE, MACHINE_VIEW, ServiceModel, TriggerModelsResponse } from '@wso2/ballerina-core';
+import debounce from 'lodash.debounce';
 
 import { CardGrid, PanelViewMore, Title, TitleWrapper } from './styles';
 import { BodyText } from '../../styles';
@@ -25,14 +26,17 @@ import ButtonCard from '../../../components/ButtonCard';
 import { isBetaModule } from './componentListUtils';
 import { RelativeLoader } from '../../../components/RelativeLoader';
 import { getEntryNodeIcon } from './EventIntegrationPanel';
-import { useCentralTriggerSearch } from './useCentralTriggerSearch';
 
 interface CentralSearchPanelProps {
     /** The page-level search query; the panel is expected to be rendered only when non-empty. */
     query: string;
     /** Locally available triggers, used to hide Central results already installed. */
     triggers: TriggerModelsResponse;
+    /** Handles a pick instead of navigating to the service wizard, for a caller that stays in place. */
+    onSelect?: (model: ServiceModel, isLocalRepository: boolean) => void;
 }
+
+const SEARCH_DEBOUNCE_MS = 700;
 
 /**
  * "More on Ballerina Central" — the remote half of the artifact gallery's page-level search.
@@ -42,13 +46,79 @@ interface CentralSearchPanelProps {
  */
 export function CentralSearchPanel(props: CentralSearchPanelProps) {
     const { rpcClient } = useRpcContext();
-    const { enabled, searching, results, localRepositoryResults } = useCentralTriggerSearch(props.query);
+    const [searching, setSearching] = useState<boolean>(true);
+    const [results, setResults] = useState<ServiceModel[]>([]);
+    const [localRepositoryResults, setLocalRepositoryResults] = useState<ServiceModel[]>([]);
+    const [additionalTriggerSearchEnabled, setAdditionalTriggerSearchEnabled] = useState<boolean>(false);
 
-    if (!enabled) {
+    const isMountedRef = useRef(true);
+    // The most recently *dispatched* search query. Debounce only coalesces rapid keystrokes into
+    // one call per pause — it does nothing once two calls are genuinely in flight together (e.g. a
+    // slow response for an older query outlasting a newer one). Comparing against this before
+    // applying a response is what stops a stale result from overwriting a fresher one purely
+    // because of network timing.
+    const latestQueryRef = useRef<string>("");
+
+    useEffect(() => {
+        isMountedRef.current = true;
+        return () => {
+            isMountedRef.current = false;
+        };
+    }, []);
+
+    useEffect(() => {
+        rpcClient
+            .getCommonRpcClient()
+            .additionalTriggerSearchEnabled()
+            .then((enabled) => {
+                if (isMountedRef.current) {
+                    setAdditionalTriggerSearchEnabled(enabled);
+                }
+            });
+    }, [rpcClient]);
+
+    // Debounced so we don't hit Central on every keystroke.
+    const runSearch = useMemo(
+        () =>
+            debounce((searchQuery: string) => {
+                latestQueryRef.current = searchQuery;
+                setSearching(true);
+                rpcClient
+                    .getServiceDesignerRpcClient()
+                    .searchTriggers({ query: searchQuery, includeLocalRepository: additionalTriggerSearchEnabled })
+                    .then((res) => {
+                        if (!isMountedRef.current || latestQueryRef.current !== searchQuery) {
+                            return;
+                        }
+                        setResults(res?.local ?? []);
+                        setLocalRepositoryResults(res?.localRepositoryResults ?? []);
+                    })
+                    .finally(() => {
+                        if (isMountedRef.current && latestQueryRef.current === searchQuery) {
+                            setSearching(false);
+                        }
+                    });
+            }, SEARCH_DEBOUNCE_MS),
+        [rpcClient, additionalTriggerSearchEnabled]
+    );
+
+    useEffect(() => {
+        if (!additionalTriggerSearchEnabled) {
+            return;
+        }
+        runSearch(props.query);
+        return () => runSearch.cancel();
+    }, [props.query, runSearch, additionalTriggerSearchEnabled]);
+
+    if (!additionalTriggerSearchEnabled) {
         return null;
     }
 
     const handleSelect = async (model: ServiceModel, isLocalRepository: boolean) => {
+        if (props.onSelect) {
+            props.onSelect(model, isLocalRepository);
+            return;
+        }
         await rpcClient.getVisualizerRpcClient().openView({
             type: EVENT_TYPE.OPEN_VIEW,
             location: {
